feat: Migrate To Typescript
