--- .coveralls.yml
@@ -1,2 +1,0 @@
-service_name: travis-ci
-repo_token: 39107FxIUVLisOeYpI6KlJd0bsRoZXghx
