+++ .lintstagedrc.json
@@ -1,0 +1,5 @@
+{
+  "*.ts": [
+    "npm run lint"
+  ]
+}
