+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+    "printWidth": 150,
+    "tabWidth": 2,
+    "singleQuote": true,
+    "trailingComma": "all",
+    "semi": true,
+    "arrowParens": "avoid"
+}
