--- .travis.yml
@@ -1,6 +1,0 @@
- language: node_js
- node_js:
-   - 6
-#after_script:
-#   - "istanbul cover test/index.js --captureExceptions spec/ && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage"
-after_success: npm run coverage
