+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "request-ip",
-  "version": "3.3.0",
+  "version": "4.0.0",
   "description": "A small Node.js module to retrieve the request's IP address",
   "keywords": [
     "request ip",
@@ -38,36 +38,45 @@
     "type": "git",
     "url": "https://github.com/pbojinov/request-ip.git"
   },
-  "license": "MIT",
-  "author": "Petar Bojinov <petarbojinov+github@gmail.com>",
   "contributors": [
     {
       "name": "Jon Peck",
       "email": "jpeck@fluxsauce.com"
     }
   ],
-  "files": [
-    "lib"
-  ],
-  "main": "./lib/index.js",
+  "main": "./dis/index.js",
+  "license": "MIT",
+  "author": "Petar Bojinov <petarbojinov+github@gmail.com>",
   "scripts": {
-    "build": "babel src --out-dir lib",
-    "changelog": "github_changelog_generator -u pbojinov -p request-ip",
-    "coverage": "nyc report --reporter=text-lcov | coveralls",
-    "test": "nyc --reporter=html --reporter=text --check-coverage --lines=100 --statements=100 tape ./test/*.js"
+    "build": "swc src -d dist --source-maps --copy-files",
+    "build:tsc": "tsc && tsc-alias",
+    "test": "ts-node -r tsconfig-paths/register --transpile-only test/run.ts",
+    "lint": "eslint --ignore-path .gitignore --ext .ts src/",
+    "lint:fix": "npm run lint -- --fix"
   },
-  "prettier": "@shopify/prettier-config",
-  "devDependencies": {
-    "@babel/cli": "^7.0.0",
-    "@babel/core": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@shopify/eslint-plugin": "^41.3.0",
-    "@shopify/prettier-config": "^1.1.2",
-    "coveralls": "^3.0.2",
-    "eslint": "^8.16.0",
-    "nyc": "^13.1.0",
-    "request": "^2.54.0",
+  "dependencies": {
+    "http": "^0.0.1-security",
+    "net": "^1.0.2",
+    "request": "^2.88.2",
     "tap-spec": "^5.0.0",
-    "tape": "^4.9.1"
+    "tape": "^5.6.3"
+  },
+  "devDependencies": {
+    "@swc/cli": "^0.1.51",
+    "@swc/core": "^1.2.108",
+    "@types/node": "^16.0.1",
+    "@typescript-eslint/eslint-plugin": "^4.28.2",
+    "@typescript-eslint/parser": "^4.28.2",
+    "eslint": "^7.30.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "lint-staged": "^11.0.0",
+    "prettier": "^2.3.2",
+    "sequelize-cli": "^6.2.0",
+    "supertest": "^6.2.2",
+    "ts-node": "^10.4.0",
+    "tsc-alias": "^1.4.1",
+    "tsconfig-paths": "^3.10.1",
+    "typescript": "^4.5.2"
   }
 }
