+++ test/run.ts
@@ -1,0 +1,4 @@
+console.log('start testing request api');
+Promise.all([import('./is'), import('./index')]).then();
+
+
